fix: Ajuste en la obtención de datos para la Cotización Finalizada
- **Caso de Uso**: Se modificó el caso de uso get_finalized_quotation_use_case.py. Este ajuste se centra en la lógica para **recuperar y procesar** los datos de una cotización que ya ha sido marcada como finalizada.
- **Esquemas de la API**: Se actualizó end_quotation_schemas.py (esquemas de entrada/salida de la API) para asegurar que los datos solicitados y recibidos para la cotización finalizada sean **correctos y consistentes**.
- **Impacto**: Esto corrige errores en la visualización o el procesamiento de datos al consultar cotizaciones en su estado final, mejorando la estabilidad del módulo.

+++ app/adapters/inbound/api/schemas/end_quotation_schemas.py
@@ -21,7 +21,7 @@ class ProveedorInfoDTO(BaseModel):
 class ProductoDTO(BaseModel):
     """DTO para un único producto en la lista."""
     id: int = Field(..., description="ID del producto")
-    cant: int = Field(..., description="Cantidad del producto")
+    cant: float = Field(..., description="Cantidad del producto")
     und: str = Field(..., description="Unidad de medida")
     nombre: str = Field(..., description="Nombre del producto")
     marca: str = Field(..., description="Marca del producto")
